Qualify `freshclam --quiet`
This commit qualifies `freshclam --quiet` in order to ensure it runs without
error. Previously, the command would fail to run as the exec was unqualified,
and Puppet could not find the executable.

Puppet refuses to trust that the $PATH of the shell from which a Puppet run is
started is correct. It therefore demands qualified `exec` types.