Amend exec timeout from 300 to 480
The default value of the `timeout` parameter used in the exec type in Puppet is
300secs. freshclam takes longer than 300secs to complete at times. If this
happens when a box is brought up, the provisioning may fail. Even though
freshclam will re-run after a while once the box has booted, it is far better to
ensure that we get a successful output from freshclam during provision, rather
than an error relating to it timing out.

This value has been set to 480, or eight minutes, given that the average run
time over three or so runs (as reported in freshclam.log) was about 6mins. Using
8mins gives us time in case a run takes longer, but not long enough that we'd
wait a long time for it to fail due to an error before reporting its failure.