Amend CHANGELOG
Release 0.0.2 was released earlier today with a change to qualify the `exec`
type in the `run_freshclam.pp` manifest. A later change was made to increase the
`timeout` parameter on the `exec` type to ensure that `freshclam` completes
successfully during a Puppet run without failing. Given that both of these
changes are related to the same issue, I have rolled the latter change in with
the former.